eatin-server/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.4
chardet==3.0.4
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
googlemaps==3.1.3
idna==2.8
mysqlclient==1.4.4
pkg-resources==0.0.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
drf-jwt==1.13.2
get==2019.4.13
pipenv==2018.11.26
post==2019.4.13
public==2019.4.13
PyJWT==1.7.1
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
virtualenv==16.7.2
virtualenv-clone==0.5.3
XlsxWriter==1.2.1

